Move class Versions to internal package
